Moved results card common functions to utilities
