+++ src/Components/EditPositionDetails/PositionDetailsCard/PositionDetailsCard.jsx
@@ -6,9 +6,8 @@ import { COMMON_PROPERTIES } from 'Constants/EndpointParams';
 import { Row } from 'Components/Layout';
 import DefinitionList from 'Components/DefinitionList';
 import InteractiveElement from 'Components/InteractiveElement';
-import { getDifferentials, getResult } from 'Components/ResultsCard/ResultsCard';
 import LanguageList from 'Components/LanguageList';
-import { getPostName, propOrDefault } from 'utilities';
+import { getDifferentials, getPostName, getResult, propOrDefault } from 'utilities';
 import {
   NO_BUREAU, NO_DATE, NO_GRADE,
   NO_POSITION_NUMBER, NO_POST, NO_SKILL, NO_TOUR_OF_DUTY, NO_UPDATE_DATE, NO_USER_LISTED,
